Player movement now works with ES6
