+++ package.json
@@ -35,7 +35,7 @@
     "expose-loader": "^0.7.5",
     "html-webpack-plugin": "^3.2.0",
     "raw-loader": "^0.5.1",
-    "webpack": "^4.18.0",
+    "webpack": "^4.25.1",
     "webpack-cli": "^3.1.0"
   }
 }
