+++ webpack.config.js
@@ -4,15 +4,14 @@ var HtmlWebpackPlugin = require('html-webpack-plugin')
 var BrowserSyncPlugin = require('browser-sync-webpack-plugin')
 var CopyWebpackPlugin = require('copy-webpack-plugin');
 
-// Phaser webpack config
-var phaserModule = path.join(__dirname, '/node_modules/phaser/')
-var phaser = path.join(phaserModule, 'src/phaser.js')
-
 var definePlugin = new webpack.DefinePlugin({
     __DEV__: JSON.stringify(JSON.parse(process.env.BUILD_DEV || 'true')),
-    WEBGL_RENDERER: true, // I did this to make webpack work, but I'm not really sure it should always be true
-    CANVAS_RENDERER: true // I did this to make webpack work, but I'm not really sure it should always be true
+    WEBGL_RENDERER: true,
+    CANVAS_RENDERER: true 
 })
+
+var phaserModule = path.join(__dirname, '/node_modules/phaser/')
+var phaser = path.join(phaserModule, 'src/phaser.js')
 
 module.exports = {
     entry: {
@@ -22,6 +21,13 @@ module.exports = {
         vendor: ['phaser']
     },
     devtool: 'cheap-source-map',
+    module: {
+        rules: [
+            { test: /\.js$/, use: ['babel-loader'], include: path.join(__dirname, 'src') },
+            { test: /phaser-split\.js$/, use: ['expose-loader?Phaser'] },
+            { test: [/\.vert$/, /\.frag$/], use: 'raw-loader' }
+        ]
+    },
     output: {
       pathinfo: true,
         path: path.resolve(__dirname, 'dev'),
@@ -33,22 +39,10 @@ module.exports = {
     watch: true, 
     plugins: [
         definePlugin,
-        //new webpack.optimize.CommonsChunkPlugin({ name: 'vendor'/* chunkName= */, filename: 'vendor.bundle.js'/* filename= */ }),
         new HtmlWebpackPlugin({
             filename: '../index.html',
             chunks: ['vendor', 'app'],
             chunksSortMode: 'manual',
-            minify: {
-                removeAttributeQuotes: false,
-                collapseWhitespace: false,
-                html5: false,
-                minifyCSS: false,
-                minifyJS: false,
-                minifyURLs: false,
-                removeComments: false,
-                removeEmptyAttributes: false
-            },
-            hash: false
         }),
         new BrowserSyncPlugin({
             host: process.env.IP || 'localhost',
@@ -58,21 +52,6 @@ module.exports = {
             }
         })
     ],
-    module: {
-        rules: [
-            { test: /\.js$/, use: ['babel-loader'], include: path.join(__dirname, 'src') },
-            { test: /phaser-split\.js$/, use: ['expose-loader?Phaser'] },
-            { test: [/\.vert$/, /\.frag$/], use: 'raw-loader' }
-        ]
-    },
-   /* node: {
-        fs: 'empty',
-        net: 'empty',
-        tls: 'empty'
-    },
-    resolve: {
-        alias: {
-            'phaser': phaser,
-        }
-    }*/
+    
+   
 }
